pyproject.toml
[tool.poetry]
name = "nixhashsync"
version = "0.1.0"
description = "nixhashsync"
authors = ["luis <ldesiqueira@pinger.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
nixhashsync = "nixhashsync.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
